feat：根据moa票据校验用户权限并返回安全信息 #9795
