+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/BkHttpRequestService.kt
@@ -1,0 +1,72 @@
+package com.tencent.devops.auth.service
+
+import com.fasterxml.jackson.core.type.TypeReference
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.tencent.devops.auth.pojo.ResponseDTO
+import com.tencent.devops.common.api.exception.RemoteServiceException
+import com.tencent.devops.common.api.util.OkhttpUtils
+import okhttp3.MediaType.Companion.toMediaTypeOrNull
+import okhttp3.Request
+import okhttp3.RequestBody.Companion.toRequestBody
+import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.stereotype.Service
+
+@Service
+class BkHttpRequestService @Autowired constructor(
+    val objectMapper: ObjectMapper
+) {
+    @Value("\${auth.appCode:}")
+    private val appCode = ""
+
+    @Value("\${auth.appSecret:}")
+    private val appSecret = ""
+    fun <T> executeHttpPost(url: String, body: Any): ResponseDTO<T> {
+        val headerStr = objectMapper.writeValueAsString(mapOf("bk_app_code" to appCode, "bk_app_secret" to appSecret))
+            .replace("\\s".toRegex(), "")
+        val requestBody = objectMapper.writeValueAsString(body)
+            .toRequestBody("application/json; charset=utf-8".toMediaTypeOrNull())
+
+        val request = Request.Builder()
+            .url(url)
+            .post(requestBody)
+            .addHeader("x-bkapi-authorization", headerStr)
+            .build()
+        return executeHttpRequest(url, request)
+    }
+
+    fun <T> executeHttpGet(url: String): ResponseDTO<T> {
+        val headerStr = objectMapper.writeValueAsString(mapOf("bk_app_code" to appCode, "bk_app_secret" to appSecret))
+            .replace("\\s".toRegex(), "")
+        val request = Request.Builder()
+            .url(url)
+            .addHeader("x-bkapi-authorization", headerStr)
+            .get()
+            .build()
+        return executeHttpRequest(url, request)
+    }
+
+    fun <T> executeHttpRequest(url: String, request: Request): ResponseDTO<T> {
+        OkhttpUtils.doHttp(request).use {
+            if (!it.isSuccessful) {
+                logger.warn("request failed, uri:($url)|response: ($it)")
+                throw RemoteServiceException("request failed, response:($it)")
+            }
+            val responseStr = it.body!!.string()
+            val responseDTO: ResponseDTO<T> =
+                objectMapper.readValue(responseStr, object : TypeReference<ResponseDTO<T>>() {})
+            if (responseDTO.code != 0L || !responseDTO.result) {
+                // 请求错误
+                logger.warn("request failed, url:($url)|response:($it)")
+                throw RemoteServiceException("request failed, response:(${responseDTO.message})")
+            }
+            logger.info("request response：${objectMapper.writeValueAsString(responseDTO.data)}")
+            return responseDTO
+        }
+    }
+
+    companion object {
+        private val logger = LoggerFactory.getLogger(BkHttpRequestService::class.java)
+    }
+}
